# backend/core/policy_engine/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Grandway Policy-Registered API",
    "version": "1.0.0",
    "description": "Generated from the Core Policy Engine registry (source of truth for endpoints, permission keys, HTTP methods, and risk levels). Request/response body schemas are documented per app in `docs/DATA_CONTRACT.md`."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/journeys/{journey_id}/stage/": {
      "post": {
        "operationId": "applicant_journeys.journey.change_stage",
        "summary": "Change Journey Stage",
        "description": "Move a journey between active stages. Cannot set completed, closed, or deferred.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "applicant_journeys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "applicant_journeys.journey.change_stage",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/leads/{lead_id}/convert/": {
      "post": {
        "operationId": "leads.lead.convert",
        "summary": "Convert Lead to Applicant",
        "description": "Create an applicant and an initial journey from the lead, link both permanently, and move the lead to its terminal converted stage. Admin only; idempotent.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "leads.lead.convert",
        "x-risk-level": "critical",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/leads/{lead_id}/follow-up/": {
      "post": {
        "operationId": "leads.lead.record_followup",
        "summary": "Record Lead Follow-up",
        "description": "Record that manual follow-up occurred, optionally with a note and a stage change.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "leads.lead.record_followup",
        "x-risk-level": "low",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/leads/{lead_id}/lost/": {
      "post": {
        "operationId": "leads.lead.mark_lost",
        "summary": "Mark Lead Lost",
        "description": "Close a lead as lost with a mandatory reason. Nothing is deleted.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "leads.lead.mark_lost",
        "x-risk-level": "high",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/leads/{lead_id}/stage/": {
      "post": {
        "operationId": "leads.lead.change_stage",
        "summary": "Change Lead Stage",
        "description": "Move a lead between active stages. Cannot set lost or converted.\n\nRequest/response field schemas: see the owning app's `docs/DATA_CONTRACT.md`.",
        "tags": [
          "leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-permission-key": "leads.lead.change_stage",
        "x-risk-level": "medium",
        "x-operation-type": "custom",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "description": "Resource payload — see the owning app's DATA_CONTRACT.md."
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "APP_RESOURCE_REASON code."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "Meta": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "next": {
            "type": [
              "string",
              "null"
            ]
          },
          "previous": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Standard error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}